pull name search results into seprate client comopnent
